Add test for set_run_args and set_batch_args (lsf)
