+++ tests/full_wlm/test_launch_orc_lsf_batch.py
@@ -109,22 +109,3 @@ def test_launch_lsf_cluster_orc_reconnect(fileutils, wlmutils):
             exp_2.stop(reloaded_orc)
             assert False
     exp_2.stop(reloaded_orc)
-
-
-def test_orc_converter_lsf_batch():
-    def converter(host):
-        int_dict = {"host1": "HOST1-IB", "host2": "HOST2-IB"}
-        if host in int_dict.keys():
-            return int_dict[host]
-        else:
-            return ""
-
-    orc = LSFOrchestrator(
-        6780,
-        db_nodes=3,
-        batch=True,
-        hosts=["batch", "host1", "host2"],
-        hostname_converter=converter,
-    )
-    assert orc.entities[0].hosts == ["HOST1-IB", "HOST2-IB"]
-    assert orc.batch_settings.batch_args["m"] == '"batch host1 host2"'
